Preparation for migrating preferences
