+++ Source/ui_android/java/com/virtualapplications/play/Constants.java
@@ -4,4 +4,19 @@ public class Constants
 {
 	public static String TAG = "Play!";
 	public static final int READ_WRITE_PERMISSION = 0;
+
+	/**
+	 * Preferences
+	 */
+	public static final String PREF_EMU_GENERAL_SHOWFPS = "ui.showfps";
+	public static final String PREF_EMU_GENERAL_SHOWVIRTUALPAD = "ui.showvirtualpad";
+	public static final String PREF_EMU_VIDEO_RESFACTOR = "renderer.opengl.resfactor";
+	public static final String PREF_EMU_VIDEO_PRESENTATIONMODE = "renderer.presentationmode";
+	public static final String PREF_EMU_AUDIO_BUFFERSIZE = "audio.spublockcount";
+
+	public static final String PREF_UI_CATEGORY_STORAGE = "ui.storage";
+	public static final String PREF_UI_RESCAN = "ui.rescan";
+	public static final String PREF_UI_CLEAR_UNAVAILABLE = "ui.clear_unavailable";
+	public static final String PREF_UI_CLEAR_CACHE = "ui.clearcache";
+	public static final String PREF_UI_THEME_SELECTION = "ui.theme_selection";
 }
