--- build_android/src/main/res/xml/settings_emu_fragment.xml
@@ -1,68 +1,0 @@
-<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android">
-	<PreferenceCategory
-		android:title="@string/pref_emu_general"
-		>
-		<CheckBoxPreference
-			android:key="ui.showfps"
-			android:title="@string/pref_emu_general_fps_display_title"
-			android:summary="@string/pref_emu_general_fps_display_summary"
-			android:persistent="false"
-		/>
-		<CheckBoxPreference
-			android:key="ui.showvirtualpad"
-			android:title="@string/pref_emu_general_pad_display_title"
-			android:summary="@string/pref_emu_general_pad_display_summary"
-			android:persistent="false"
-		/>
-	</PreferenceCategory>
-	<PreferenceCategory
-		android:title="@string/pref_emu_video"
-		>
-		<ListPreference
-			android:key="renderer.opengl.resfactor"
-			android:title="@string/pref_emu_video_res_factor_title"
-			android:entries="@array/pref_emu_video_res_factor_entries"
-			android:entryValues="@array/pref_emu_video_res_factor_values"
-			android:defaultValue="1"
-			android:negativeButtonText="@null"
-			android:positiveButtonText="@null"/>
-		<CheckBoxPreference
-			android:key="renderer.widescreen"
-			android:title="@string/pref_emu_video_widescreen_title"
-			android:summary="@string/pref_emu_video_widescreen_summary"
-			android:persistent="false"
-		/>
-		<CheckBoxPreference
-			android:key="renderer.opengl.forcebilineartextures"
-			android:title="@string/pref_emu_video_force_bilinear_title"
-			android:summary="@string/pref_emu_video_force_bilinear_summary"
-			android:persistent="false"
-		/>
-		<ListPreference
-			android:key="renderer.presentationmode"
-			android:title="@string/pref_emu_video_presentation_mode_title"
-			android:entries="@array/pref_emu_video_presentation_mode_entries"
-			android:entryValues="@array/pref_emu_video_presentation_mode_values"
-			android:defaultValue="1"
-			android:negativeButtonText="@null"
-			android:positiveButtonText="@null"/>
-	</PreferenceCategory>
-	<PreferenceCategory
-		android:title="@string/pref_emu_audio"
-		>
-		<CheckBoxPreference
-			android:key="audio.enableoutput"
-			android:title="@string/pref_emu_audio_enable_output_title"
-			android:summary="@string/pref_emu_audio_enable_output_summary"
-			android:persistent="false"
-		/>
-		<ListPreference
-			android:key="audio.spublockcount"
-			android:title="@string/pref_emu_audio_spublockcount_title"
-			android:entries="@array/pref_emu_audio_spublockcount_entries"
-			android:entryValues="@array/pref_emu_audio_spublockcount_values"
-			android:defaultValue="400"
-			android:negativeButtonText="@null"
-			android:positiveButtonText="@null"/>
-	</PreferenceCategory>
-</PreferenceScreen>
