--- build_android/src/main/res/xml/settings_headers.xml
@@ -1,10 +1,0 @@
-<preference-headers xmlns:android="http://schemas.android.com/apk/res/android">
-	<header
-		android:fragment="com.virtualapplications.play.SettingsActivity$EmulatorSettingsFragment"
-		android:title="@string/preferences_screen_emu"
-	/>
-	<header
-		android:fragment="com.virtualapplications.play.SettingsActivity$UISettingsFragment"
-		android:title="@string/preferences_screen_ui"
-    />
-</preference-headers>
