--- build_android/src/main/res/xml/settings_ui_fragment.xml
@@ -1,36 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android">
-    
-    <PreferenceCategory
-        android:key="ui.storage"
-        android:title="@string/preferences_ui">
-        <Preference
-            android:key="ui.rescan"
-            android:title="@string/pref_ui_rescan_all_title"
-            android:summary="@string/pref_ui_rescan_all_summary"
-            android:defaultValue="0"
-            android:persistent="false" />
-        <Preference
-            android:key="ui.clear_unavailable"
-            android:title="@string/pref_ui_clear_unavailable_title"
-            android:summary="@string/pref_ui_clear_unavailable_summary"
-            android:persistent="false" />
-        <Preference
-            android:key="ui.clearcache"
-            android:title="@string/pref_ui_reset_cache_title"
-            android:summary="@string/pref_ui_reset_cache_summary"
-            android:persistent="false" />
-    </PreferenceCategory>
-
-    <PreferenceCategory
-        android:title="@string/preferences_theme" >
-    <ListPreference
-        android:key="ui.theme_selection"
-        android:title="@string/pref_title_theme_selection"
-        android:entries="@array/pref_theme_selection_titles"
-        android:entryValues="@array/pref_theme_selection_values"
-        android:defaultValue="1"
-        android:negativeButtonText="@null"
-        android:positiveButtonText="@null" />
-    </PreferenceCategory>
-</PreferenceScreen>
